requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
inference-sdk  # Replace with the actual version of your SDK
# stuff  # Replace 'stuff' with the actual package name or remove if it's a custom module.
python-multipart==0.0.5